requirements.txt
opencv-python
scikit-image
scipy
requests
pillow
openai
numpy
matplotlib
jupyterlab
torch
timm
lightning